refactor: Delete unused code
