+++ app/src/main/java/org/sopt/collaboration/campuspick/feature/aftersearch/AfterSearchContract.kt
@@ -1,6 +1,5 @@
 package org.sopt.collaboration.campuspick.feature.aftersearch
 
-import okhttp3.internal.immutableListOf
 import org.sopt.collaboration.campuspick.core.ui.base.SideEffect
 import org.sopt.collaboration.campuspick.core.ui.base.UiState
 import org.sopt.collaboration.campuspick.domain.model.FilteredClub
@@ -16,7 +15,6 @@ data class AfterSearchState(
         region = null,
         clubDay = null
     ),
-    val filteredClub: List<FilteredClub> = immutableListOf(),
     val clubLoadState: ClubLoadState = ClubLoadState.Loading
 ) : UiState
